backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
python-dotenv
PyPDF2
python-docx
requests
Werkzeug
gunicorn
